Updated schema sizes and query categorization, CBO flags
Based on a recent series of runs on a arbitrary benchmark cluster. The
intent is for the queries to complete in a reasonable timeframe (in terms
of benchmarking): 15-60 seconds. The specific datasets are chosen with
that intent.

Also added CBO flag join_distribution_type.

(based on commit b238a3de47f811ca881ea4bf7db77d99399ae1ff)